extract campaign evidence routes
